pyproject.toml
[project]
name = "texture-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "reflex>=0.8.4",
]
